+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Moved and renamed `Classes.CLASS_FILE_EXTENSION` constant to `JavaClassFile.EXTENSION`.
 - Added `maven-properties-plugin` to `pom.xml` to output all Maven properties to `target/maven.properties`.
 - Added `Methods.getOneDeclared` which just delegates to `Methods.Safe.getOneDeclared`.
+- Fixed `GenericClass` accepting non generic classes as generic type arguments.
 
 ---